requirements.txt
discord.py==1.5.1
eventemitter==0.2.0
nest-asyncio==1.4.2
requests==2.24.0
